refactor: booleanGenerator 패지키 변경
