+++ src/main/java/domain/RandomBooleanGenerator.java
@@ -1,4 +1,4 @@
-package domain.numbergenerator;
+package domain;
 
 import java.util.Random;
